# openapi.yaml
openapi: 3.0.1
info:
  title: Laine Déco API
  version: 1.0.0
  description: |
    Documentation complète de l'API Laine Déco - Plateforme de vente de laine et accessoires

    ## 🔐 Authentification
    Utilisez un token JWT Bearer dans l'en-tête `Authorization: Bearer <token>`

    ## 📋 Sections de l'API
    - **Public** : Routes accessibles sans authentification
    - **Client** : Routes pour les clients authentifiés
    - **Admin** : Routes pour les administrateurs
    - **Finance** : Routes pour l'équipe finance

  contact:
    name: Support Laine Déco
    email: support@laine-deco.cm
    url: https://laine-deco.cm
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Serveur local
  - url: https://api.laine-deco.cm
    description: Serveur de production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Entrez le token JWT
  schemas:
...
    ProductResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
        category:
          type: string
        brand:
          type: string
        stock:
          type: integer
        rating:
          type: number
        createdAt:
          type: string
          format: date-time
    OrderResponse:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        status:
          type: string
        totalAmount:
          type: number
        items:
          type: array
          items:
            type: object
        createdAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
...
  /api/public/products/{id}:
    get:
      tags:
        - Public - Products
      summary: Détails du produit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Détails du produit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Produit non trouvé
...
  /api/public/promotions/product/{productId}:
    get:
      tags:
        - Public - Promotions
      summary: Promotions pour un produit
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Promotions du produit

  # ========== PUBLIC PRODUCT PACKS ROUTES ==========
...
  /api/public/product-packs/product/{productId}:
    get:
      tags:
        - Public - Product Packs
      summary: Packs contenant un produit
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Packs du produit

  # ========== PUBLIC ORDER ROUTES ==========
  /api/public/orders:
    post:
      tags:
        - Public - Orders
      summary: Créer une commande (invité)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Commande créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
...
  /api/public/maintenance-guides/product/{productId}:
    get:
      tags:
        - Public - Maintenance Guides
      summary: Guides pour un produit
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Guides du produit
...
  /api/admin/products:
    post:
      tags:
        - Admin - Products
      summary: Créer un produit
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductResponse'
      responses:
        '201':
          description: Produit créé

    get:
      tags:
        - Admin - Products
      summary: Lister les produits
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Produits
  /api/admin/products/{id}:
    get:
      tags:
        - Admin - Products
      summary: Détails du produit
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Produit

    put:
      tags:
        - Admin - Products
      summary: Mettre à jour le produit
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductResponse'
      responses:
        '200':
          description: Produit mis à jour
  /api/admin/products/search:
    get:
      tags:
        - Admin - Products
      summary: Rechercher des produits
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Résultats
...
  /api/finance/products/{id}:
    get:
      tags:
        - Finance - Products
      summary: Détails du produit
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Produit

    put:
      tags:
        - Finance - Products
      summary: Mettre à jour le produit
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductResponse'
      responses:
        '200':
          description: Produit mis à jour
  /api/finance/products/search:
    get:
      tags:
        - Finance - Products
      summary: Rechercher des produits
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Résultats
...